requirements.txt
streamlit>=1.32.0
groq>=0.9.0
textblob>=0.18.0
python-dotenv>=1.0.0
chromadb>=0.4.0
requests>=2.31.0
